requirements.txt
Flask==2.0.3
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
Flask_SocketIO==4.3.1
Flask_SQLAlchemy==2.5.1
google_auth_oauthlib==0.4.2
pddiktipy==1.0.0
PyJWT==2.7.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
Requests==2.31.0
twilio==6.38.1
